Define all versions in single location
This commit introduces a Versions object that is referenced by all
build scripts that reference versions of dependencies, plugins, and
tools. The Versions object is located in the buildSrc subproject which
is picked up by Gradle and visible in all build scripts. Plugin
versions are applied using the pluginManagement block in
settings.gradle.kts as the plugins DSL does not currently support
variables.