Convert Test Kit to a fluent API
This commit includes a major overhaul of the existing Test Kit API in
order to shift to a fluent API for all parts of the Test Kit except
ExecutionEventConditions and TestExecutionResultConditions.

The key abstractions that enable the fluent API include the Events
façade, EventStatistics, and the Executions façade.

Consult the changes to DisabledTests and IgnoreConditionTests to see
examples of some of the major improvements over the previous API.

Documentation of the Test Kit in the Release Notes and User Guide will
be addressed in #1621.

Issue: #1662