Set context class loader parallel execution mode
Prior to this commit the `ForkJoinPoolHierarchicalTestExecutorService`
used the default `ForkJoinWorkerThreadFactory` which in turn sets the
context class loader to `ClassLoader.getSystemClassLoader()`.
In runtime environments that create and use a tailored context
class loader this leads to `ClassNotFoundException` when only the
`ClassLoader.getSystemClassLoader()` is available.

Now the `ForkJoinPoolHierarchicalTestExecutorService` creates
`ForkJoinWorkerThread` instances that set the context class loader
to the one available at creation time of the service.

Fixes #1646