Include stack trace of all exceptions in Assertions.assertAll()
Prior to this commit, exceptions tracked by Assertions.assertAll() were
properly reported via a MultipleFailuresError; however, the stack
traces for individual exceptions were not retained.

This commit addresses this issue by storing all such exceptions as
suppressed exceptions within the resulting MultipleFailuresError.
Consequently, the stack traces for such exceptions are now visible as
"Suppressed" at the end of the stack trace for the invocation of
assertAll().

Closes: #1602