Introduce Try type to preserve failure cause
Exceptions reported due to failed reflective operations such as loading
a class, reading a field's value, or looking up a method by name now
include the original exception as their cause to make it easier to debug
underlying issues.

Such operations in `ReflectionUtils` and `ReflectionSupport` no longer
return an instance of `Optional` but one of a new `Try` type which
encapsulates either a successful invocation with a (nullable) resulting
value or a failure with an exception as its cause. Depending on the
caller, the cause can then be logged or rethrown as appropriate.

Resolves #602.