Support multidimensional arrays as arguments with @MethodSource
Prior to this commit it was not possible for a @MethodSource factory
method to provide a stream of multidimensional arrays to a
@ParameterizedTest without explicitly wrapping the multidimensional
array in an Arguments instance or in an Object[].

This commit fixes this by passing any multidimensional array "as is" to
the parameterized test method.

In addition, this commit documents the special behavior for arguments
supplied as a one-dimensional object array (e.g., Object[], String[],
etc.). See the revised class-level Javadoc for @MethodSource for
details.

Issue: #1665