Ensure Random ordering with custom seed works with parallel execution
This commit introduces a new constraint on the order in which the
orderMethods() and getDefaultExecutionMode() methods in a MethodOrderer
are invoked.

Specifically, orderMethods() is now guaranteed to be invoked before
getDefaultExecutionMode() so that the latter may include logic that
determines the appropriate return value based on what was performed in
orderMethods().

This commit also fixes a bug in the Random MethodOrderer when a custom
seed is configured and tests are executed in parallel. Specifically,
the getDefaultExecutionMode() in the Random orderer now returns
SAME_THREAD if a custom seed has been configured and otherwise
CONCURRENT.

Issue: #13