Propagate exceptions thrown while accessing static fields
Prior to this commit, any exception thrown while accessing a static
field was silently swallowed. Consequently, errors occurring due to the
initialization of a static @RegisterExtension field in JUnit Jupiter
were not propagated to the user.

This commit ensures that all exceptions thrown while accessing a static
field are propagated up the call stack, potentially all the way to the
end user.

Fixes: #1666