Prevent thread interrupt from causing subsequent test failures
Prior to this commit, if a container or test that was executed via a
HierarchicalTestExecutorService (e.g., a @Test method in JUnit Jupiter)
interrupted the current thread (e.g., via
Thread.currentThread().interrupt()) but failed to clear the
'interrupted status' flag for the current thread, that particular
container or test may have passed but inadvertently caused subsequent
containers or tests to fail if the latter attempted to interact with
the thread in any manner that might result in an InterruptedException.

This commit fixes this issue by clearing the 'interrupted status' flag
for the current thread after each container or test executed via
HierarchicalTestExecutorService.

In addition, if the 'interrupted status' flag for the current thread is
not properly cleared by user code (e.g., the test author), a message is
now logged at DEBUG level informing the user that he or she may wish to
take further actions to ensure that the 'interrupted status' flag is
properly cleared.

Fixes: #1688