Use HTTPS to link to Maven Central
Using an HTTP connection to an artifact server can leave you vulnerable to MITM attack. Linked below is a POC created back in the days when Maven Central didn't support HTTPS that dynamically re-writes jar files as they are downloaded by maven.
https://max.computer/blog/how-to-take-over-the-computer-of-any-java-or-clojure-or-scala-developer/

The README isn't currently really advising people to use this url directly, but still, best to be safe.