Fix using resource locks on test templates
Prior to this commit, it was not possible to use `@ResourceLock` on
`@TestTemplate` methods because `TestTemplateInvocationTestDescriptor`
returned a non-empty set of `ExclusiveResources` which is disallowed for
dynamically added `TestDescriptors`. Now, an empty set is returned
instead which works because the parent `TestTemplateTestDescriptor`
returns the resources from the `@ResourceLock` annotation.

Fixes #1697.