Avoid NullPointerException when creating Executions from filtered events
Prior to this commit, the logic in Executions.createExecutions() could
result in a NullPointerException if the executions were being created
from filtered events.

For example, an attempt to create executions from
`testEvents.failed().executions()` resulted in a NullPointerException
since the "started" event (needed to determine the start Instant) was
no longer present in the event list.

This commit fixes this bug by avoiding the creation of Execution
instances if the "started" event is absent.

Consequently, an invocation such as
`testEvents.failed().executions().count()` now returns `0` instead of
throwing a NullPointerException.

Issue: #1356