Restore display names for methods
In #1719, the legacy reporting name of Vintage's `TestDescriptor` was
changed to use the fully-qualified class name, if it can be read from
the corresponding `Description`. However, if a `Description` represents
a method, the class name is also present. Thus, it was also used as
legacy reporting name for methods. This commit restores the legacy
reporting names for methods but keeps using the FQCN for classes.