Support @TempDir field injection by implementing BeforeEachCallback
This commit improves the support for @TempDir field injection by
performing the injection as a BeforeEachCallback instead of a
TestInstancePostProcessor.

Kudos to @marcphilipp for the suggestion.

This change allows the @TempDir field to be re-injected with different
temp directories between test method executions, thereby enabling the
use of different temp directories between test methods in the same test
class.

Issue: #1738