Fix test coverage for :junit-jupiter-params
Prior to this commit the coverage for `junit-jupiter-params` was missing
from the aggregated JaCoCo report because the tests were executed using
different class files than those used by downstream projects that use
`@ParameterizedTests`. The classes are different because we relocate
classes using the shadow plugin which modifies the bytecode. The code
in the buildscript that attempted to replace the classes directory with
the shadowed jar file for the project's `test` task was broken since the
migration to Kotlin DSL in 82eb832.

Note to self: `classpath.minus(files)` does not modify `classpath` but
returns a new one.