Use default toString() when reporting competing ParameterResolvers
Prior to this commit, if multiple competing ParameterResolvers of the
exact same type were encountered, the error message included multiple
occurrences of the same fully qualified class name which did not
provide any information about the instances.

This commit improves the error message in such cases by including the
default toString() output for each ParameterResolver. This ensures that
the type and hash code are reported for each resolver even if the
resolver implementation overrides toString().