Supply expected and actual values for failed boolean assertions
Prior this commit, the `assertTrue` and `assertFalse` methods in
`Assertions` did not supply the actual and expected boolean values to
the `AssertionFailedError` constructor. This is inconsistent with other
assertions and can cause problems for tools that rely on the presence of
the actual and expected values -- such as Eclipse's support for
displaying multiple failures in a pop-up window.

Resolves #1781.