Support searches for nested repeatable annotations
Prior to this commit, the search algorithm for repeatable annotations
did not find repeatable annotations if they were used as
meta-annotations on other repeatable annotations.

This commit updates the search algorithm in
AnnotationUtils.findRepeatableAnnotations() to find repeatable
annotations on repeatable annotations. This functionality is
transitively available in AnnotationSupport.findRepeatableAnnotations().

Prerequisite for: #1793

Closes: #1849