azure-pipelines: more complete port of travis conf
A more complete port of the Travis configuration.

Instead of using the gradle task, run it on the command line.
This allows us to use the same script configuration for container-based
builds and on-agent builds.

For newer versions of java (12 and 13) use a container to perform the
build.

Add code coverage builds, and publishing jobs when appropriate.