Report intermediate containers finished lazily
Since JUnit 4 does report events for intermediate containers, the
Vintage engine has to synthesize them. Until now this was done lazily
for start events, i.e. when the first child was started, but eagerly for
finish events, i.e. when the last static child of a container was
finished. However, the latter approach is problematic for containers
that contain both static and dynamic children, for example a Spock
specification with a regular and an unrolled feature method. Instead,
intermediate containers are now reported as finished as soon as a test
is started for which the container is not an ancestor.

Fixes #1819.