Add assertLinesMatch overloads taking a custom message
Prior to this commit, `Assertions.assertLinesMatch` was the only
assertion that didn't support user-supplied error messages. This
commit adds overloads that take a String or Supplier<String> as
a custom message. That message is prepended to the automatically
generated detail message.

This commit also improves the textual layout of the automatically
generated detail message by emitting each expected and actual
line into a dedicated new line.

Closes #1855

Co-Authored-By: Sam Brannen <sbrannen@pivotal.io>