Introduce declarative timeouts
This commit introduces declarative timeouts for testable and lifecycle
methods. The API surface consists of a new `@Timeout` annotation that
can be applied to `@Test`, `@TestTemplate`,` @TestFactory`, and
lifecycle methods. In addition, it can be applied to a test class which
has the same meaning as declaring it on all of its testable, but not its
lifecycle methods.

Moreover, default timeouts can be defined using configuration parameters
for all of the above methods, all lifecycle methods, all testable
methods, or each method type individually.

The timeout is implemented in a preemptive manner, but without having to
execute the test in a different thread. Instead, a single-thread
`ExecutorService` is used to schedule an interrupt of the original test
thread. This way, `ThreadLocal` will continue to work which is important
for integration tests that use frameworks like Spring and its
`@Transactional` support.

Resolves #80.