Relax @RegisterExtension field type restriction
Prior to this commit, the declared field type for an extension
registered via @RegisterExtension was required to implement an
extension API (or at least the Extension marker interface). This put an
unnecessary burden on the extension developer and simultaneously
exposed JUnit Jupiter extension APIs to users of the extension. This
restriction effectively does not allow for proper separation of concerns
or encapsulation.

This commit addresses this issue by moving the "instance of Extension"
check to field evaluation time. Consequently, a @RegisterExtension
field may now be declared as any type implemented by the evaluated
field value.

Closes #1908