Improve diagnostics for private @RegisterExtension fields
Prior to this commit, if a @RegisterExtension field was declared as
private, JUnit Jupiter silently ignored it. Consequently, it was
difficult for a user to determine why the extension "was not working".

This commit improves the diagnostics in such scenarios by failing the
corresponding test method or test class if a @RegisterExtension field
is private.

Closes #1912