Simplify nohttp plugin configuration
Thanks to a tip from @sdeleuze, this commit simplifies the `nohttp`
plugin configuration by replacing the original hack with a simpler one.

Original hack: introduce a dummy whitelist file, configure it in the
`nohttp` configuration block, and introduce exclusions in the
`suppressions.xml` file.

New hack: cast the `source` to a `ConfigurableFileTree` in the `nohttp`
configuration block and use it as intended.

See #1923, https://github.com/spring-io/nohttp/issues/21