Add a requirement field to AbstractTestDescriptor
The AbstractTestDescriptor class was extended with a field to hold the
requirement, including the implementation of a get()-method.
Another constructor was added so the requirement can be set. The other
constructors were changed to call this new constructor with null-value,
as it was before where the constructor without a source called the one
with a null-value for source.
The equals-method was not changed as the comparison is done by the
uniqueId.

Issue: #1900