Fix discovery of inherited nested classes
The discovery algorithm introduced in 5.5 remembers already the
resolution of each discovery selector. Prior to this commit, nested
classes were resolved by adding a ClassSelector for the nested class.
If that ClassSelector has been resolved before, discovery is assumed to
be finished. Similarly, when resolving MethodSelectors for methods
within a nested test class that have been resolved before, discovery
will stop.

This commit introduces dedicated selectors for nested classes and their
methods which take into account the actual enclosing classes that will
be present at runtime. These selectors are package-private for now but
could be made publicly available in a future release.

Fixes #1954.