Test discovery of dynamic Vintage test class by unique ID
A completely dynamic test class, i.e. a test class with only dynamic
children, such as a Spock spec with only `@Unroll` feature methods,
must not be filtered with the unique ID of the test class. This was
broken in 5.4.x and before and now works due to the reworked of test
discovery in 5.5. This commit adds a regression test to ensure this
edge case continues to work.