Avoid JupiterTestEngine crash if Hamcrest is not on the classpath
Prior to this commit, the JupiterTestEngine would crash without
executing any tests if JUnit 4 was on the classpath but Hamcrest was
not.

This commit fixes this bug by ensuring that initialization of the
OpenTest4JAndJUnit4AwareThrowableCollector class no longer fails if the
org.junit.internal.AssumptionViolatedException class cannot be loaded
from the classpath -- for example, due to a missing Hamcrest dependency.

Fixes #2004