Cache methods to improve performance of TestSource creation
Prior to this commit, we scanned the hierarchy of a test class for
methods once for every test method. This lead to quadratic runtime
behavior and was problematic for large test classes. Now, the new
`TestSourceProvider` caches the methods of a class so that we only need
to filter that list once for every test method.

Resolves #1965.