Improve handling of duplicate Descriptions
Some runners don't memoize Description instances and even pass different
instances when invoking testStarted and testFinished. This makes it
particularly difficult when there are multiple equal Description
instances in a Runner.

Prior to this commit, such cases were deemed unresolvable which lead to
multiple dynamic TestDescriptors with the same unique ID being reported
which is a violation of the contract of TestExecutionListener.

Now, we assume the testStarted/testSkipped event for a Description with
multiple equal Descriptions is referring to the first Description that
hasn't yet been started or skipped. The corresponding TestDescriptor is
then reported as started/skipped. When testFinished is called for such
a Description, we check whether the same thread has previously reported
an equal Description as started. If so, we report the corresponding
TestDescriptor as finished.

Fixes #2021.