Added generation of additional javadoc for jupiter-api only
The documentation build generates separate javadoc that contains only
API now. The new docs will be published to github pages under
/docs/current/jupiter-api.

This was done by duplicating the previous javadoc goal, with no
optimizations. I removed the asciidoc dependency because I saw now use
for it in the code and building locally works fine without it.

Resolves Issue: #676