Revert build to use Java 11
We rely on tools such as Gradle, JaCoCo, Asciidoctor, and others and the
past few months have shown that's it's unrealistic to expect all of
them to be compatible with the latest JDK as soon as it's released. For
example, we're currently stuck on JDK 12 because Asciidoctor does not
yet support JDK 13. Since we don't want to compromise the safety of our
contributors, we should not rely on outdated non-LTS releases. Instead,
we now switch back the build to be compatible with the JDK 11 LTS
release which will get security patches until JDK 17 is released.

Closes #2018. Closes #2061. Closes #1906.