Detect cycles in inner class hierarchies
Prior to this commit, an invocation of
ReflectionSupport.findNestedClasses() could result in infinite
recursion if an inner class hierarchy formed a cycle via inheritance.

This commit addresses this by introducing cycle detection in
ReflectionUtils.findNestedClasses.

Consequently, the detection of cycles in @Nested test class hierarchies
will cause the JUnit Jupiter test engine to halt with an exception
instead of infinitely recursing.

Closes: #2039