Log at DEBUG level when JUnit 4 is on classpath but Hamcrest is not
Prior to this commit, JUnit Jupiter emitted an INFO level log message
to inform users when JUnit 4 is present on the classpath but Hamcrest
is not. This log message is useful if one intends to use JUnit 4
assumptions; however, since Java Util Logging (JUL) logs INFO messages
to STD_ERR, these messages showed up as RED in many IDEs which gave the
impression of an error.

In order to improve usability for use cases when Hamcrest is
intentionally not on the classpath but JUnit 4 is, this commit updates
OpenTest4JAndJUnit4AwareThrowableCollector to consistently log only at
DEBUG level.

Closes #2098