Improve documentation for lifecycle callback extension APIs
Prior to this commit, users expressed confusion regarding the semantics
of callback APIs such as AfterEachCallback vs. AfterTestExecutionCallback
with regard to wrapping of user-defined setup and teardown methods.

This commit attempts to address this by clarifying the semantics in the
first paragraph of the corresponding class-level Javadoc for such APIs.