Download element-list for AssertJ to fix Javadoc links
The package-list no longer exists in the version of the Javadocs we now
link to. Moreover, since all external repos now have element-list files
the build script is now a little simpler. To avoid issues that stem from
URLs not having a trailing slash in the future, the configuration is now
checked for it being present.

Resolves #2200.