Replace "blacklisted" with "unrecoverable" exceptions
This commit replaces the oppressive term "blacklist" [1] with a more
descriptive, domain-specific one. Exceptions that are rethrown by the
new `UnrecoverableExceptions.rethrowIfUnrecoverable` method, are
_unrecoverable_, i.e. they should always terminate test plan execution
immediately instead of being caught and potentially wrapped in another
exception.

To support running newer engines on an older Platform, the existing
class is deprecated instead of deleted for now.

[1] https://tools.ietf.org/id/draft-knodel-terminology-00.html#rfc.section.1.2