Introduce MethodName orderer to replace deprecated Alphanumeric
With the introduction of DisplayName orderer in 6fee44d, the name of the
Alphanumeric orderer becomes ambiguous ie "Alphanumeric of what?".
MethodName is a better name. It is unambiguous, it clearly states what
it is using for ordering the tests and it's naming is consistent with
the new orderer DisplayName.

Resolves: #2334