Add overload of DynamicTest.stream() that accepts a Stream
Prior to this commit, the DynamicTest.stream() factory method accepted
an Iterator as the input source; however, it is inconvenient to use
this factory method when working directly with streams for the input
source.

This commit introduces an overloaded variant of DynamicTest.stream()
that accepts a Stream to address such use cases.

Closes #2374