Let Testable annotation target any element type
Prior to this commit the usage of `@Testable` was restricted to
declarations of types, methods, and fields. As custom test engines
may use different element types as their "testable constructs",
this restriction is lifted in backward and future-compatible manner.
By dropping the `@Target` annotation, the default behaviour as
defined by the Java Language Specification applies:

  If an annotation of type java.lang.annotation.Target is not present
  on the declaration of an annotation type T, then T is applicable in
  all declaration contexts except type parameter declarations, and in
  no type contexts.

https://docs.oracle.com/javase/specs/jls/se8/html/jls-9.html#jls-9.6.4.1

Closes #2391