Evaluate argument length only when argument is not null
JUnit Jupiter 5.7.0 introduced a regression in conjunction with the
work performed to address #2358. Specifically, an attempt to truncate
an argument to a parameterized test results in a NullPointerException
if the toString() implementation of the given argument returns null.

This regression has already been addressed indirectly in #2410;
however, we would like to keep the additional non-null check in
ParameterizedTestNameFormatter. In addition, we want to keep the test
introduced in PR #2408 as a regression test.

Closes #2408
Fixes #2405