Return "null" from StringUtils.nullSafeToString if toString() returns null
Prior to this commit, the implementation of StringUtils.nullSafeToString
did not align with the Javadoc. Specifically, the documentation states
that nullSafeToString() never returns null; however, if the invocation
of toString() on the supplied object returned null, nullSafeToString()
simply returned that null value.

This commit fixes this by ensuring that nullSafeToString() returns "null"
if the invocation of toString() returns null.

Closes #2410