Only call creators that throw exceptions once
Prior to this commit, creator functions passed to ExtensionContext.Store
were called repeatedly if the first invocation threw an exception. Now,
they are only called once and the exception will be rethrown whenever
the value stored for that key is read. However, the exception will not
be rethrown when closing the store because only successfully created
`CloseableResources` can be closed.

Resolves #2455.