Fix concurrency unit tests on some platforms
This commit _attempts_ to fix concurrency unit test failures observed on
Windows Server 2019 10.0 amd64 with JVM 15.0.1 (Azul Systems, Inc.
15.0.1+8) and Linux 5.4.0-1031-azure amd64 with JVM 15 (Oracle
Corporation 15+36-1562). It does this by using a `CyclicBarrier` to
ensure test methods run concurrently when expected to. This is the same
logic used by `ParallelExecutionIntegrationTests`.

Note, I was not able to reproduce the build failures locally.

Issue: #2229