Remove spurious usage of the word simply
The word simply is used in roughly two situations:

 * When used in instructional documentation the word simply often conveys a
   value judgement about the complexity of a task. Not only is this
   unnecessary, it makes a number of assumptions about the readers familiarity
   with the concepts involved and can come across as condescending.

 * When used to describe a default implementation, the word simply is used to
   to convey that the description is a complete description of the
   implementation. This should be unnecessary given the context.

As such nearly all usages of the world simply can be removed without a loss of
clarity.

For more context and a longer argument I found the arguments made Jim Fisher in
his talk "Don't Say Simply" at Write the Docs Prague 2018[1] to be quite
compelling.

 1. https://www.youtube.com/watch?v=gsT2BBWBVmM