Improve failure message for assertNull()
Prior to this commit, if the `toString()` implementation for the
`actual` object passed to assertNull() returned `null` or `"null"`, the
following assertion failure message was generated, which is very
confusing and difficult to debug.

    expected <null> but was <null>

This commit improves the assertion failure message in such cases by
including the actual object's type and identity hash in addition to the
result of invoking `toString()` (i.e., "null").

For example, invoking `assertNull("null")` will now result in an
assertion failure message similar to the following.

    expected: null<null> but was: java.lang.String@264b3504<null>

Closes #2523