Improve failure message for null vs. null comparisons
Prior to this commit, if the expected and actual values for a failed
assertion both had "null" string representations and if either the
expected or actual value was the null reference, the generated failure
message looked something like the following.

    expected: null<null> but was: java.lang.String@264b3504<null>

This commit improves the assertion failure message in such cases by
representing the null reference as "<null>" instead of "null<null>",
leading to generated failure messages similar to the following.

    expected: <null> but was: java.lang.String@264b3504<null>

This improvement applies to assertEquals(), assertNull, and any other
assertion that internally delegates to format(Object, Object, String) or
formatValues(Object, Object) in AssertionUtils.

See #2523