Introduce AssertionTestUtils.assertMessageMatches
assertMessageMatches() is a convenience method for using a RegEx instead
of assertMessageStartsWith()/assertMessageEndsWith() combos.

This commit makes use of the new method in AssertNullAssertionsTests but
leaves all other existing tests unchanged.

assertMessageMatches() can be used for new tests, or we can slowly
migrate to it where it makes sense as we modify existing tests for some
other reason.