Fix container failure XML reporting (#2542)
Prior to this commit, failing containers were only reported in case they
contained at least one test. However, for example for a parameterized
Jupiter tests and an exception in a `@BeforeAll` method, that led to
failures being silently swallowed. Now, in addition to tests, leaf
nodes of the test tree are always included in the XML report, even if
they are containers, not tests.

Moreover, failures on the container level that occurred after their
children had been executed were not reported, e.g. when an exception
was thrown from a Jupiter `@AfterAll` method. Now such failures cause
the contained tests to be reported as failed.

Fixes #2537.