Named arguments in parameterized tests (#2521)
This commit introduces a `Named` interface, which is used to wrap test arguments and give them names.
Thanks to this, arguments in parameterized tests with `@MethodSource` or `@ArgumentSource` can now have optional names. When the argument is included in the display name of an iteration, this name will be used instead of the value.