Remove package name validation from classpath scanning (#2571)
The validation was problematic since it re-implemented name validation
to avoid a dependency to a JDK class and was missing checks for
primitive types. Removing it is fine since classpath scanning with
invalid package names does not cause errors but merely doesn't find any
classes, just like classpath scanning for a non-existent package.

Fixes #2570.