Suggested fix for currently disabled test in ReflectionUtilsWithGenericTypeHierarchiesTests.
If accepted, this commit will address part of the outstanding work in issue 981.
It addresses findsMethodsIndependentlyFromOrderOfImplementationsOfInterfaces by
ensuring a consistent ordering of Class<?> objects when getting the implemented
interfaces for a type. This is handled by a new method, ReflectionUtils.getInterfaces.
I wanted a consistent ordering and arbitrarily went for alphanumeric sorting
by implemented interface name.

I also went further than was necessary to fix the failing test
by ensuring that all calls to clazz.getInterfaces() in ReflectionUtils
used this method. I did this for consistency within that class.

Issue: #981