Support concurrent use cases in the UniqueIdTrackingListener
Prior to this commit, the UniqueIdTrackingListener wrote to an output
file with a specific name. This approach works fine when only one
listener is writing the file per test suite. However, when multiple
TestPlans are executed in parallel for the same test suite (for
example, with parallel test execution in a Gradle build), each
UniqueIdTrackingListener will attempt to recreate and write to the
exact same file. The result is that the "last one wins", meaning that
the output file only contains the tracked UIDs for the last
UniqueIdTrackingListener, and all other tracked UIDs are lost.

This commit fixes this by ensuring that each UniqueIdTrackingListener
instance writes to an output file that includes a random number in the
file name, thereby creating one unique output file per listener.

Consequently, consumers of the output files must now read the UIDs from
all such output files. In addition, the output files must be deleted
prior to TestPlan execution since the UniqueIdTrackingListener no
longer overwrites the previous output file.

Closes #2619