Improve logging for extension registration in JUnit Jupiter
This commit improves the logging for extension registration within
JUnit Jupiter by introducing the following categories for logged
extensions.

- default
- auto-detected
- local

All registered extensions are now logged consistently, and aggregated
logging for default and auto-detected extensions has been removed.

If the `source` is a Method or Field, the source information is
prefixed with "method" or "field", respectively, and the source is
converted to a short description of the method or field.

Furthermore, the `source` is no longer logged if it does not provide
additional context.

Closes #2679