Disallow duplicate registration via @RegisterExtension and @ExtendWith
Prior to this commit, @ExtendWith and @RegisterExtension could be
declared on a field to register duplicate extensions of the same time,
but that scenario likely does not make sense.

This commit ensures that @RegisterExtension and @ExtendWith cannot be
used to register an extension of the same type for a given field, by
throwing a PreconditionViolationException if the user attempts to do so.

See #2680