Support conversion of null source value in TypedArgumentConverter
Prior to this commit, if a concrete subclass of TypedArgumentConverter
was supplied a `null` source value, the convert(Object, ParameterContext)
method in TypedArgumentConverter threw a NullPointerException.

This commit ensures that TypedArgumentConverter supports the conversion
of a `null` source value in parameterized tests by delegating
immediately to the protected convert(S) method with a value of `null`,
thereby avoiding a NullPointerException.

Note, however, that concrete subclasses of TypedArgumentConverter must
properly support the conversion of `null` into the configured target
type.

Fixes #2682