Ensure TypedArgumentConverter can convert to a primitive
Prior to this commit, a TypedArgumentConverter would throw an
ArgumentConversionException when attempting to convert from a wrapper
type to its corresponding primitive type (for example, from Integer to
int), even though Java reflection APIs permit such an invocation for a
method.

This commit fixes this by checking for compatibility using the new
ReflectionUtils.isAssignableTo(Class,Class) method introduced in
commit a0dd37160c81167ac05376888debb64b00315dc9.

This commit also improves the error messages for source and target
compatibility failures in TypedArgumentConverter.

Fixes: #2685