Add suppressed exceptions when creating MultipleFailuresError in Vintage
Prior to this commit, if a JUnit 4 based test threw an
org.junit.runners.model.MultipleFailureException that contained
multiple failures, the TestRun class in the VintageTestEngine created
an org.opentest4j.MultipleFailuresError that contained the multiple
exceptions as "failures", but it did not contain those exceptions as
suppressed exceptions. Consequently, the stack traces for those
exceptions were lost, making it harder for users to deduce the cause of
the test failure.

This commit fixes this by adding all exceptions as suppressed
exceptions when creating a MultipleFailuresError.

Fixes #2693