Support @Nested test class constructors with arguments on JDK 8 again
The recently added support for registering extensions via @ExtendWith
on parameters introduced a regression for @Nested test classes whose
constructors accept arguments. Namely, a PreconditionViolationException
is now thrown stating the following.

Parameter index must not be 0 for an inner class constructor

Note, however, that this regression only applies to @Nested test classes
compiled with JDK 8, due to a bug in javac in JDK 8.

This commit fixes this regression by effectively not searching for
annotations on the parameter for the implicit enclosing instance when
the JDK 8 bug is detected.

Fixes: #2704